+++ Source/Mastering/AssemblyMetadata.cs
@@ -18,10 +18,10 @@
 //
 
 #if !DISABLE_ASSEMBLY_VERSION_ATTRIBUTE
-[assembly: AssemblyVersion("2019.3.3.16010")]
+[assembly: AssemblyVersion("2019.3.4.16010")]
 #endif
-[assembly: AssemblyFileVersion("2019.3.3.16010")]
-[assembly: AssemblyInformationalVersion("2019.3.3 (Beta build 2019.3.3.16010)")]
+[assembly: AssemblyFileVersion("2019.3.4.16010")]
+[assembly: AssemblyInformationalVersion("2019.3.4 (Beta build 2019.3.4.16010)")]
 
 #endif
 
@@ -34,9 +34,9 @@ namespace Gapotchenko.FX
     {
         public const string ProductName = "Gapotchenko.FX";
         public const string CompanyName = "Gapotchenko";
-        public const string InformationalVersion = "2019.3.3 (Beta build 2019.3.3.16010)";
-        public const string DisplayVersion = "2019.3.3";
-        public const string DisplayVersion2F = "2019.3";
+        public const string InformationalVersion = "2019.3.4 (Beta build 2019.3.4.16010)";
+        public const string DisplayVersion = "2019.3.4";
+        public const string DisplayVersion2F = "2019.4";
     }
 #endif
 }